Add files via upload
VowelCountFinderApp

Given the supplied text file (INPUT.TXT), write a Java program that loads all words and determines the average number of vowels per word grouped by: set of vowels present in a word and length of the word. Result should be written to the output file (OUTPUT.TXT).

Assumptions:

 Capital and lower case letter should be treated as the same.
 Input contains only English words and punctuation, words are separated by at least one whitespace character.
 Input is small enough to fit in memory.
 If any additional assumptions about input\output are made, they should be stated in comments/email.
Example:

INPUT

Platon made bamboo boats.

OUTPUT

({a, o}, 6) -> 2.5 ({a, o}, 5) -> 2 ({a, e}, 4) -> 2

Rules:

  This exercise should be treated as a regular work task (i.e. apply usual professional dev practices: OOP, SOLID, unit tests, project structure).
  The candidate must work independently, no collaboration is permitted.
  The candidate may conduct research either online or offline as they see fit, in order to complete the task. Please do not copy a solution from Internet.
  On completion, the resulting files must be sent with Reply to All.
Task intended to be done in a 4 hours. Please send us version which you count as done with
Please note, that structure of the code .